requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
python-dotenv==1.0.1
tenacity==9.0.0
PyYAML==6.0.2
